+++ src/components/ClasesCliente/SeccionClases.jsx
@@ -13,7 +13,7 @@ const SeccionClases = () => {
         <div className="clase-card">
           <h3>Agendar nueva clase</h3>
           <p>Reserva una clase disponible según tu horario y preferencia.</p>
-          <button onClick={() => navigate("/cliente/clases/agendar")}>
+          <button onClick={() => navigate("/cliente/agendar")}>
             Agendar
           </button>
         </div>
@@ -21,7 +21,7 @@ const SeccionClases = () => {
         <div className="clase-card">
           <h3>Ver mis clases</h3>
           <p>Consulta las clases que ya tienes reservadas.</p>
-          <button onClick={() => navigate("/cliente/clases/mis-clases")}>
+          <button onClick={() => navigate("/cliente/mis-clases")}>
             Ver clases
           </button>
         </div>
@@ -29,7 +29,7 @@ const SeccionClases = () => {
         <div className="clase-card">
           <h3>Ver historial</h3>
           <p>Mira las clases que ya tomaste o cancelaste.</p>
-          <button onClick={() => navigate("/cliente/clases/historial")}>
+          <button onClick={() => navigate("/cliente/historial")}>
             Ver historial
           </button>
         </div>
@@ -37,7 +37,7 @@ const SeccionClases = () => {
         <div className="clase-card">
           <h3>Ver horario de clases</h3>
           <p>Consulta los horarios disponibles en el calendario semanal.</p>
-          <button onClick={() => navigate("/cliente/clases/horario")}>
+          <button onClick={() => navigate("/cliente/horario")}>
             Ver horario
           </button>
         </div>
