+++ src/styles/Administrador/DashboardAdmin.css
@@ -1,3 +1,4 @@
+
 .admin-layout {
     display: flex;
     min-height: 100vh;
