+++ src/styles/ClasesCliente/AgendarClase.css
@@ -3,7 +3,6 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  min-height: 100vh; /* Para empujar el footer hacia abajo */
 }
 
 .card-agendar {
