+++ src/styles/Entrenador/DashboardEntrenador.css
@@ -215,4 +215,10 @@ html, body {
   }
 }
 
+.dashboard-entrenador{
+  width: 90%;
+  margin-left: auto;
+  margin-right: auto;
+}
+
