+++ src/styles/InicioSesionFN/Footer.css
@@ -4,6 +4,8 @@
   text-align: center;
   padding: 1rem 0;
   width: 100%;
+  position: fixed;
+  bottom: 0px;
 }
 
 .social-icons {
